Update the jwtproxy to use role assignments for team membership.
No-Issue
